fix: corrigir erros de compilação no Limiter e Phase1Activity
- Limiter: abs(Short) não existe em kotlin.math — converter para Int antes:
  abs(futureL.toInt()). O erro de compareTo era cascata deste tipo não resolvido.
- Phase1Activity: when não exaustivo após adição de AMBIENT_TOO_QUIET no enum
  HardwareValidationResult — adicionar branch com mensagem orientando o usuário.

+++ app/src/main/java/com/showraw/android/audio/Limiter.kt
@@ -52,7 +52,7 @@ class Limiter(
             // Amostra futura (lookahead) — usada para calcular gain necessário agora
             val futureL = buffer[i]
             val futureR = buffer[i + 1]
-            val level   = max(abs(futureL), abs(futureR)) / 32768f
+            val level   = max(abs(futureL.toInt()), abs(futureR.toInt())) / 32768f
 
             gainReduction = if (level > thr) {
                 val target = thr / level
